agent/pyproject.toml
[project]
name = "ats-agent"
version = "1.0.0"
description = "自动化测试管理平台Agent节点"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "websockets>=10.0",
    "psutil>=5.8.0",
    "pyyaml>=5.4.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "mypy>=1.7.1",
]

# 注意：由于main是async函数，不能直接作为entry point
# 可以使用: uv run python agent.py 或 uv run agent/agent.py
# [project.scripts]
# ats-agent = "agent.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# 指定要包含的Python包（目录）
# 由于代码文件直接在agent目录下，使用packages配置
packages = ["."]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

